Bugfix on setup module.
