+++ triage/config.ini
@@ -1,11 +1,36 @@
 [paths]
-# Make sure workDir is either empty or does not exist.
-# This is the directory to store intermediate results.
-workDir = 
-# This is the path to fuzzbench experiment data folder.
-fuzzbenchExpDir = 
-# This is the directory to store triage output.
-outDir = 
+# Store raw data and intermediate results.
+workDir = path/to/workDir
+# Stores triage outputs.
+outDir = path/to/outDir
+# Path to FuzzBench experiment-data folder.
+fuzzbenchExpDir = path/to/experiment-data
+# List of Fuzzbench experiment names.
+# A benchmark-fuzzer pair must only appear in one experiment.
+# All fuzzers running a benchmark must only appear in the same experiment.
+[experiments]
+exp1
+exp2
+# RevBugBench benchmark names.
+# RevbugBench currently support the following 10 benchmarks.
+[benchmarks]
+binutils-fuzz_cxxfilt
+curl
+libpcap
+libxml2_xml
+usrsctp
+binutils-fuzz_disassemble
+lcms
+libxml2_reader
+proj4
+zstd
+# Fuzzer names.
+[fuzzers]
+afl
+aflplusplus
+eclipser
+fairfuzz
+libfuzzer
 [values]
-# The number of current tasks to run in parellel.
-cores = 
+# The number of triage tasks to run in parellel.
+cores = 4
